+++ components/FeatureHighlight/tests/snapshot/MDCFeatureHighlightSnapshotTests.m
@@ -17,10 +17,26 @@
 
 #import <XCTest/XCTest.h>
 
+/**
+ An MDCFeatureHighlightViewController subclass that allows the user to override the @c
+ traitCollection property.
+ */
+@interface MDCFeatureHighlightViewControllerWithCustomTraitCollection
+    : MDCFeatureHighlightViewController
+@property(nonatomic, strong) UITraitCollection *traitCollectionOverride;
+@end
+
+@implementation MDCFeatureHighlightViewControllerWithCustomTraitCollection
+- (UITraitCollection *)traitCollection {
+  return self.traitCollectionOverride ?: [super traitCollection];
+}
+@end
+
 /** Snapshot tests for MDCFeatureHighlightViewController and MDCFeatureHighlightView. */
 @interface MDCFeatureHighlightSnapshotTests : MDCSnapshotTestCase
 
-@property(nonatomic, strong) MDCFeatureHighlightViewController *featureHighlightViewController;
+@property(nonatomic, strong)
+    MDCFeatureHighlightViewControllerWithCustomTraitCollection *featureHighlightViewController;
 
 @end
 
@@ -59,8 +75,9 @@
   [currentViewController.view addSubview:highlightLabel];
   highlightLabel.center = currentViewController.view.center;
   self.featureHighlightViewController =
-      [[MDCFeatureHighlightViewController alloc] initWithHighlightedView:highlightLabel
-                                                              completion:nil];
+      [[MDCFeatureHighlightViewControllerWithCustomTraitCollection alloc]
+          initWithHighlightedView:highlightLabel
+                       completion:nil];
   self.featureHighlightViewController.titleText = @"Feature Highlight Title";
   self.featureHighlightViewController.bodyText = @"Feature Highlight Body";
   self.featureHighlightViewController.outerHighlightColor = UIColor.blueColor;
@@ -80,6 +97,94 @@
   [self snapshotVerifyViewForIOS13:window];
 
   [highlightLabel removeFromSuperview];
+}
+
+- (void)testPreferredFontForAXXXLContentSizeCategory {
+  if (@available(iOS 11.0, *)) {
+    // Given
+    UIWindow *window = [[[UIApplication sharedApplication] delegate] window];
+    UIViewController *currentViewController = window.rootViewController;
+    UILabel *highlightLabel = [[UILabel alloc] init];
+    highlightLabel.text = @"Highlight";
+    [highlightLabel sizeToFit];
+    [currentViewController.view addSubview:highlightLabel];
+    highlightLabel.center = currentViewController.view.center;
+    self.featureHighlightViewController =
+        [[MDCFeatureHighlightViewControllerWithCustomTraitCollection alloc]
+            initWithHighlightedView:highlightLabel
+                         completion:nil];
+    self.featureHighlightViewController.titleText = @"Feature Highlight Title";
+    self.featureHighlightViewController.bodyText = @"Feature Highlight Body";
+    UITraitCollection *xsTraitCollection = [UITraitCollection
+        traitCollectionWithPreferredContentSizeCategory:UIContentSizeCategoryExtraSmall];
+    self.featureHighlightViewController.traitCollectionOverride = xsTraitCollection;
+    UIFontMetrics *bodyMetrics = [UIFontMetrics metricsForTextStyle:UIFontTextStyleBody];
+    UIFont *originalFont = [bodyMetrics scaledFontForFont:[UIFont fontWithName:@"Zapfino" size:12]];
+    self.featureHighlightViewController.titleFont = originalFont;
+    self.featureHighlightViewController.bodyFont = originalFont;
+    self.featureHighlightViewController.adjustsFontForContentSizeCategory = YES;
+
+    // When
+    UITraitCollection *aXXXLTraitCollection =
+        [UITraitCollection traitCollectionWithPreferredContentSizeCategory:
+                               UIContentSizeCategoryAccessibilityExtraExtraExtraLarge];
+    self.featureHighlightViewController.traitCollectionOverride = aXXXLTraitCollection;
+    XCTestExpectation *expectation =
+        [[XCTestExpectation alloc] initWithDescription:@"Feature highlight is presented"];
+    [currentViewController presentViewController:self.featureHighlightViewController
+                                        animated:YES
+                                      completion:^{
+                                        [expectation fulfill];
+                                      }];
+
+    // Then
+    [self waitForExpectations:@[ expectation ] timeout:5];
+    [self snapshotVerifyView:window];
+  }
+}
+
+- (void)testPreferredFontForXSContentSizeCategory {
+  if (@available(iOS 11.0, *)) {
+    // Given
+    UIWindow *window = [[[UIApplication sharedApplication] delegate] window];
+    UIViewController *currentViewController = window.rootViewController;
+    UILabel *highlightLabel = [[UILabel alloc] init];
+    highlightLabel.text = @"Highlight";
+    [highlightLabel sizeToFit];
+    [currentViewController.view addSubview:highlightLabel];
+    highlightLabel.center = currentViewController.view.center;
+    self.featureHighlightViewController =
+        [[MDCFeatureHighlightViewControllerWithCustomTraitCollection alloc]
+            initWithHighlightedView:highlightLabel
+                         completion:nil];
+    self.featureHighlightViewController.titleText = @"Feature Highlight Title";
+    self.featureHighlightViewController.bodyText = @"Feature Highlight Body";
+    UITraitCollection *aXXXLTraitCollection =
+        [UITraitCollection traitCollectionWithPreferredContentSizeCategory:
+                               UIContentSizeCategoryAccessibilityExtraExtraExtraLarge];
+    self.featureHighlightViewController.traitCollectionOverride = aXXXLTraitCollection;
+    UIFontMetrics *bodyMetrics = [UIFontMetrics metricsForTextStyle:UIFontTextStyleBody];
+    UIFont *originalFont = [bodyMetrics scaledFontForFont:[UIFont fontWithName:@"Zapfino" size:12]];
+    self.featureHighlightViewController.titleFont = originalFont;
+    self.featureHighlightViewController.bodyFont = originalFont;
+    self.featureHighlightViewController.adjustsFontForContentSizeCategory = YES;
+
+    // When
+    UITraitCollection *xsTraitCollection = [UITraitCollection
+        traitCollectionWithPreferredContentSizeCategory:UIContentSizeCategoryExtraSmall];
+    self.featureHighlightViewController.traitCollectionOverride = xsTraitCollection;
+    XCTestExpectation *expectation =
+        [[XCTestExpectation alloc] initWithDescription:@"Feature highlight is presented"];
+    [currentViewController presentViewController:self.featureHighlightViewController
+                                        animated:YES
+                                      completion:^{
+                                        [expectation fulfill];
+                                      }];
+
+    // Then
+    [self waitForExpectations:@[ expectation ] timeout:5];
+    [self snapshotVerifyView:window];
+  }
 }
 
 @end