+++ components/TextFields/tests/unit/ContainedInputView/MDCBaseTextFieldTests.m
@@ -184,7 +184,7 @@
   [textField layoutIfNeeded];
 
   // Then
-  CGRect expectedClearButtonFrame = CGRectMake(99, 27, 19, 19);
+  CGRect expectedClearButtonFrame = CGRectMake(99, 22, 19, 19);
   CGRect actualClearButtonFrame = [textField clearButtonRectForBounds:textFieldFrame];
   XCTAssertTrue(CGRectEqualToRect(actualClearButtonFrame, expectedClearButtonFrame));
 }
@@ -305,7 +305,7 @@
 
   // Then
   CGSize newSize = textField.frame.size;
-  CGSize correctSize = CGSizeMake(130, 57);
+  CGSize correctSize = CGSizeMake(130, 63);
   XCTAssertTrue(CGSizeEqualToSize(newSize, correctSize));
 }